+++ app/src/main/java/com/danilkinkin/buckwheat/di/AppModule.kt
@@ -33,4 +33,8 @@ object AppModule {
     @Singleton
     @Provides
     fun provideStorageDao(db: DatabaseModule) = db.storageDao()
+
+    @Singleton
+    @Provides
+    fun provideSavedTagDao(db: DatabaseModule) = db.savedTagDao()
 }